fcp-microservices/services/workbench-services/task_status_update/requirements.txt
psycopg2-binary
requests
python-dotenv
boto3